+++ examples/basic_stealth.py
@@ -2,10 +2,10 @@
 
 from playwright.sync_api import sync_playwright
 
-from pw_stealth import stealth_context
+from pw_stealth import stealth_context_sync
 
 with sync_playwright() as p:
-    context = stealth_context(p, headless=False)
+    context = stealth_context_sync(p, headless=False)
     page = context.new_page()
     page.goto("https://bot.sannysoft.com", wait_until="networkidle")
     page.screenshot(path="sannysoft.png", full_page=True)